Allow only administrator to delete all comments, allow user to uupvote/downvote each comment exactly once

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteServlet.java
@@ -14,7 +14,8 @@
 
 package com.google.sps.servlets;
 
-
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 import com.google.cloud.datastore.Datastore;
 import com.google.cloud.datastore.DatastoreOptions;
 import com.google.cloud.datastore.Entity;
@@ -37,13 +38,16 @@ public class DeleteServlet extends HttpServlet {
    */
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-    Datastore datastore = DatastoreOptions.getDefaultInstance().getService();
-    Query<Key> query =
-      Query.newKeyQueryBuilder()
-        .setKind("Comment")
-        .build();
-    Key[] keys = Iterators.toArray(datastore.run(query), Key.class);
-    datastore.delete(keys);
+    UserService userService = UserServiceFactory.getUserService();
+    if (userService.isUserAdmin()) {
+      Datastore datastore = DatastoreOptions.getDefaultInstance().getService();
+      Query<Key> query =
+        Query.newKeyQueryBuilder()
+          .setKind("Comment")
+          .build();
+      Key[] keys = Iterators.toArray(datastore.run(query), Key.class);
+      datastore.delete(keys);
+    }
   }
 }
